Move library2 to original package
