+++ app/src/main/java/eu/kanade/tachiyomi/ui/library/LibraryHolder.kt
@@ -1,0 +1,28 @@
+package eu.kanade.tachiyomi.ui.library
+
+import android.view.View
+import eu.davidea.flexibleadapter.FlexibleAdapter
+import eu.davidea.viewholders.FlexibleViewHolder
+import eu.kanade.tachiyomi.data.database.models.Manga
+
+/**
+ * Generic class used to hold the displayed data of a manga in the library.
+ * @param view the inflated view for this holder.
+ * @param adapter the adapter handling this holder.
+ * @param listener a listener to react to the single tap and long tap events.
+ */
+
+abstract class LibraryHolder(
+        view: View,
+        adapter: FlexibleAdapter<*>
+) : FlexibleViewHolder(view, adapter) {
+
+    /**
+     * Method called from [LibraryCategoryAdapter.onBindViewHolder]. It updates the data for this
+     * holder with the given manga.
+     *
+     * @param manga the manga to bind.
+     */
+    abstract fun onSetValues(manga: Manga)
+
+}
