+++ app/src/main/java/eu/kanade/tachiyomi/ui/library/LibraryMangaEvent.kt
@@ -1,7 +1,7 @@
 package eu.kanade.tachiyomi.ui.library
 
 import eu.kanade.tachiyomi.data.database.models.Category
-import eu.kanade.tachiyomi.ui.library2.LibraryItem
+import eu.kanade.tachiyomi.ui.library.LibraryItem
 
 class LibraryMangaEvent(val mangas: Map<Int, List<LibraryItem>>) {
