--- app/src/main/java/eu/kanade/tachiyomi/ui/library2/LibraryPresenter.kt
@@ -1,314 +1,0 @@
-package eu.kanade.tachiyomi.ui.library2
-
-import android.os.Bundle
-import android.util.Pair
-import com.hippo.unifile.UniFile
-import com.jakewharton.rxrelay.BehaviorRelay
-import eu.kanade.tachiyomi.data.cache.CoverCache
-import eu.kanade.tachiyomi.data.database.DatabaseHelper
-import eu.kanade.tachiyomi.data.database.models.Category
-import eu.kanade.tachiyomi.data.database.models.Manga
-import eu.kanade.tachiyomi.data.database.models.MangaCategory
-import eu.kanade.tachiyomi.data.download.DownloadManager
-import eu.kanade.tachiyomi.data.preference.PreferencesHelper
-import eu.kanade.tachiyomi.data.preference.getOrDefault
-import eu.kanade.tachiyomi.source.LocalSource
-import eu.kanade.tachiyomi.source.SourceManager
-import eu.kanade.tachiyomi.source.online.HttpSource
-import eu.kanade.tachiyomi.ui.base.presenter.BasePresenter
-import eu.kanade.tachiyomi.ui.library.LibrarySort
-import eu.kanade.tachiyomi.util.combineLatest
-import eu.kanade.tachiyomi.util.isNullOrUnsubscribed
-import rx.Observable
-import rx.Subscription
-import rx.android.schedulers.AndroidSchedulers
-import rx.schedulers.Schedulers
-import uy.kohesive.injekt.Injekt
-import uy.kohesive.injekt.api.get
-import java.io.IOException
-import java.io.InputStream
-import java.util.*
-
-/**
- * Presenter of [LibraryController].
- */
-class LibraryPresenter(
-        private val db: DatabaseHelper = Injekt.get(),
-        private val preferences: PreferencesHelper = Injekt.get(),
-        private val coverCache: CoverCache = Injekt.get(),
-        private val sourceManager: SourceManager = Injekt.get(),
-        private val downloadManager: DownloadManager = Injekt.get()
-) : BasePresenter<LibraryController>() {
-
-    /**
-     * Categories of the library.
-     */
-    var categories: List<Category> = emptyList()
-        private set
-
-    /**
-     * Relay used to apply the UI filters to the last emission of the library.
-     */
-    private val filterTriggerRelay = BehaviorRelay.create(Unit)
-
-    /**
-     * Relay used to apply the selected sorting method to the last emission of the library.
-     */
-    private val sortTriggerRelay = BehaviorRelay.create(Unit)
-
-    /**
-     * Library subscription.
-     */
-    private var librarySubscription: Subscription? = null
-
-    override fun onCreate(savedState: Bundle?) {
-        super.onCreate(savedState)
-        subscribeLibrary()
-    }
-
-    /**
-     * Subscribes to library if needed.
-     */
-    fun subscribeLibrary() {
-        if (librarySubscription.isNullOrUnsubscribed()) {
-            librarySubscription = getLibraryObservable()
-                    .combineLatest(filterTriggerRelay.observeOn(Schedulers.io()),
-                            { lib, _ -> Pair(lib.first, applyFilters(lib.second)) })
-                    .combineLatest(sortTriggerRelay.observeOn(Schedulers.io()),
-                            { lib, _ -> Pair(lib.first, applySort(lib.second)) })
-                    .map { Pair(it.first, it.second.mapValues { it.value.map(::LibraryItem) }) }
-                    .observeOn(AndroidSchedulers.mainThread())
-                    .subscribeLatestCache({ view, pair ->
-                        view.onNextLibraryUpdate(pair.first, pair.second)
-                    })
-        }
-    }
-
-    /**
-     * Applies library filters to the given map of manga.
-     *
-     * @param map the map to filter.
-     */
-    private fun applyFilters(map: Map<Int, List<Manga>>): Map<Int, List<Manga>> {
-        // Cached list of downloaded manga directories given a source id.
-        val mangaDirsForSource = mutableMapOf<Long, Map<String?, UniFile>>()
-
-        // Cached list of downloaded chapter directories for a manga.
-        val chapterDirectories = mutableMapOf<Long, Boolean>()
-
-        val filterDownloaded = preferences.filterDownloaded().getOrDefault()
-
-        val filterUnread = preferences.filterUnread().getOrDefault()
-
-        val filterFn: (Manga) -> Boolean = f@ { manga ->
-            // Filter out manga without source.
-            val source = sourceManager.get(manga.source) ?: return@f false
-
-            // Filter when there isn't unread chapters.
-            if (filterUnread && manga.unread == 0) {
-                return@f false
-            }
-
-            // Filter when the download directory doesn't exist or is null.
-            if (filterDownloaded) {
-                // Get the directories for the source of the manga.
-                val dirsForSource = mangaDirsForSource.getOrPut(source.id) {
-                    val sourceDir = downloadManager.findSourceDir(source)
-                    sourceDir?.listFiles()?.associateBy { it.name }.orEmpty()
-                }
-
-                val mangaDirName = downloadManager.getMangaDirName(manga)
-                val mangaDir = dirsForSource[mangaDirName] ?: return@f false
-
-                val hasDirs = chapterDirectories.getOrPut(manga.id!!) {
-                    mangaDir.listFiles()?.isNotEmpty() ?: false
-                }
-                if (!hasDirs) {
-                    return@f false
-                }
-            }
-            true
-        }
-
-        return map.mapValues { entry -> entry.value.filter(filterFn) }
-    }
-
-    /**
-     * Applies library sorting to the given map of manga.
-     *
-     * @param map the map to sort.
-     */
-    private fun applySort(map: Map<Int, List<Manga>>): Map<Int, List<Manga>> {
-        val sortingMode = preferences.librarySortingMode().getOrDefault()
-
-        val lastReadManga by lazy {
-            var counter = 0
-            db.getLastReadManga().executeAsBlocking().associate { it.id!! to counter++ }
-        }
-
-        val sortFn: (Manga, Manga) -> Int = { manga1, manga2 ->
-            when (sortingMode) {
-                LibrarySort.ALPHA -> manga1.title.compareTo(manga2.title)
-                LibrarySort.LAST_READ -> {
-                    // Get index of manga, set equal to list if size unknown.
-                    val manga1LastRead = lastReadManga[manga1.id!!] ?: lastReadManga.size
-                    val manga2LastRead = lastReadManga[manga2.id!!] ?: lastReadManga.size
-                    manga1LastRead.compareTo(manga2LastRead)
-                }
-                LibrarySort.LAST_UPDATED -> manga2.last_update.compareTo(manga1.last_update)
-                LibrarySort.UNREAD -> manga1.unread.compareTo(manga2.unread)
-                else -> throw Exception("Unknown sorting mode")
-            }
-        }
-
-        val comparator = if (preferences.librarySortingAscending().getOrDefault())
-            Comparator(sortFn)
-        else
-            Collections.reverseOrder(sortFn)
-
-        return map.mapValues { entry -> entry.value.sortedWith(comparator) }
-    }
-
-    /**
-     * Get the categories and all its manga from the database.
-     *
-     * @return an observable of the categories and its manga.
-     */
-    private fun getLibraryObservable(): Observable<Pair<List<Category>, Map<Int, List<Manga>>>> {
-        return Observable.combineLatest(getCategoriesObservable(), getLibraryMangasObservable(),
-                { dbCategories, libraryManga ->
-                    val categories = if (libraryManga.containsKey(0))
-                        arrayListOf(Category.createDefault()) + dbCategories
-                    else
-                        dbCategories
-
-                    this.categories = categories
-                    Pair(categories, libraryManga)
-                })
-    }
-
-    /**
-     * Get the categories from the database.
-     *
-     * @return an observable of the categories.
-     */
-    private fun getCategoriesObservable(): Observable<List<Category>> {
-        return db.getCategories().asRxObservable()
-    }
-
-    /**
-     * Get the manga grouped by categories.
-     *
-     * @return an observable containing a map with the category id as key and a list of manga as the
-     * value.
-     */
-    private fun getLibraryMangasObservable(): Observable<Map<Int, List<Manga>>> {
-        return db.getLibraryMangas().asRxObservable()
-                .map { list -> list.groupBy { it.category } }
-    }
-
-    /**
-     * Requests the library to be filtered.
-     */
-    fun requestFilterUpdate() {
-        filterTriggerRelay.call(Unit)
-    }
-
-    /**
-     * Requests the library to be sorted.
-     */
-    fun requestSortUpdate() {
-        sortTriggerRelay.call(Unit)
-    }
-
-    /**
-     * Called when a manga is opened.
-     */
-    fun onOpenManga() {
-        // Avoid further db updates for the library when it's not needed
-        librarySubscription?.let { remove(it) }
-    }
-
-    /**
-     * Returns the common categories for the given list of manga.
-     *
-     * @param mangas the list of manga.
-     */
-    fun getCommonCategories(mangas: List<Manga>): Collection<Category> {
-        if (mangas.isEmpty()) return emptyList()
-        return mangas.toSet()
-                .map { db.getCategoriesForManga(it).executeAsBlocking() }
-                .reduce { set1: Iterable<Category>, set2 -> set1.intersect(set2) }
-    }
-
-    /**
-     * Remove the selected manga from the library.
-     *
-     * @param mangas the list of manga to delete.
-     * @param deleteChapters whether to also delete downloaded chapters.
-     */
-    fun removeMangaFromLibrary(mangas: List<Manga>, deleteChapters: Boolean) {
-        // Create a set of the list
-        val mangaToDelete = mangas.distinctBy { it.id }
-        mangaToDelete.forEach { it.favorite = false }
-
-        Observable.fromCallable { db.insertMangas(mangaToDelete).executeAsBlocking() }
-                .onErrorResumeNext { Observable.empty() }
-                .subscribeOn(Schedulers.io())
-                .subscribe()
-
-        Observable.fromCallable {
-            mangaToDelete.forEach { manga ->
-                coverCache.deleteFromCache(manga.thumbnail_url)
-                if (deleteChapters) {
-                    val source = sourceManager.get(manga.source) as? HttpSource
-                    if (source != null) {
-                        downloadManager.findMangaDir(source, manga)?.delete()
-                    }
-                }
-            }
-        }
-                .subscribeOn(Schedulers.io())
-                .subscribe()
-    }
-
-    /**
-     * Move the given list of manga to categories.
-     *
-     * @param categories the selected categories.
-     * @param mangas the list of manga to move.
-     */
-    fun moveMangasToCategories(categories: List<Category>, mangas: List<Manga>) {
-        val mc = ArrayList<MangaCategory>()
-
-        for (manga in mangas) {
-            for (cat in categories) {
-                mc.add(MangaCategory.create(manga, cat))
-            }
-        }
-
-        db.setMangaCategories(mc, mangas)
-    }
-
-    /**
-     * Update cover with local file.
-     *
-     * @param inputStream the new cover.
-     * @param manga the manga edited.
-     * @return true if the cover is updated, false otherwise
-     */
-    @Throws(IOException::class)
-    fun editCoverWithStream(inputStream: InputStream, manga: Manga): Boolean {
-        if (manga.source == LocalSource.ID) {
-            LocalSource.updateCover(context, manga, inputStream)
-            return true
-        }
-
-        if (manga.thumbnail_url != null && manga.favorite) {
-            coverCache.copyToCache(manga.thumbnail_url!!, inputStream)
-            return true
-        }
-        return false
-    }
-
-}
