+++ app/src/main/res/layout/item_library_category2.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<eu.kanade.tachiyomi.ui.library2.LibraryCategoryView
+<eu.kanade.tachiyomi.ui.library.LibraryCategoryView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
@@ -11,4 +11,4 @@
 
     </android.support.v4.widget.SwipeRefreshLayout>
 
-</eu.kanade.tachiyomi.ui.library2.LibraryCategoryView>
+</eu.kanade.tachiyomi.ui.library.LibraryCategoryView>
